# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IMPACCT Brooklyn Content Submission Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List submissions (filtered by role)
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Create a submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionBody"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /submissions/{id}:
    get:
      operationId: getSubmission
      tags: [submissions]
      summary: Get a single submission with history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /submissions/{id}/status:
    patch:
      operationId: updateSubmissionStatus
      tags: [submissions]
      summary: Update submission status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubmissionStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, department_head, staff]
        department_id:
          type: ["integer", "null"]
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - created_at
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, department_head, staff]
        department_id:
          type: ["integer", "null"]
      required:
        - name
        - email
        - password
        - role
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, department_head, staff]
        department_id:
          type: ["integer", "null"]
...
    Submission:
      type: object
      properties:
        id:
          type: integer
        department_id:
          type: integer
        submitted_by:
          type: integer
        content_type:
          type: string
          enum: [event, program_update, staff_change, resource, news]
        headline:
          type: string
        body:
          type: string
        link:
          type: ["string", "null"]
        attachment_url:
          type: ["string", "null"]
        requested_publish_date:
          type: ["string", "null"]
          format: date
        status:
          type: string
          enum: [pending, published, needs_revision]
        admin_notes:
          type: ["string", "null"]
        submitted_at:
          type: string
          format: date-time
        published_at:
          type: ["string", "null"]
          format: date-time
        submitter_name:
          type: ["string", "null"]
        department_name:
          type: ["string", "null"]
      required:
        - id
        - department_id
        - submitted_by
        - content_type
        - headline
        - body
        - status
        - submitted_at
    SubmissionDetail:
      type: object
      properties:
        id:
          type: integer
        department_id:
          type: integer
        submitted_by:
          type: integer
        content_type:
          type: string
          enum: [event, program_update, staff_change, resource, news]
        headline:
          type: string
        body:
          type: string
        link:
          type: ["string", "null"]
        attachment_url:
          type: ["string", "null"]
        requested_publish_date:
          type: ["string", "null"]
          format: date
        status:
          type: string
          enum: [pending, published, needs_revision]
        admin_notes:
          type: ["string", "null"]
        submitted_at:
          type: string
          format: date-time
        published_at:
          type: ["string", "null"]
          format: date-time
        submitter_name:
          type: ["string", "null"]
        department_name:
          type: ["string", "null"]
        history:
          type: array
          items:
            $ref: "#/components/schemas/StatusHistory"
      required:
        - id
        - department_id
        - submitted_by
        - content_type
        - headline
        - body
        - status
        - submitted_at
        - history
    StatusHistory:
      type: object
      properties:
        id:
          type: integer
        submission_id:
          type: integer
        changed_by:
          type: integer
        old_status:
          type: ["string", "null"]
        new_status:
          type: string
        note:
          type: ["string", "null"]
        changed_at:
          type: string
          format: date-time
        changed_by_name:
          type: ["string", "null"]
      required:
        - id
        - submission_id
        - changed_by
        - new_status
        - changed_at
    CreateSubmissionBody:
      type: object
      properties:
        department_id:
          type: integer
        content_type:
          type: string
          enum: [event, program_update, staff_change, resource, news]
        headline:
          type: string
        body:
          type: string
        link:
          type: ["string", "null"]
        attachment_url:
          type: ["string", "null"]
        requested_publish_date:
          type: ["string", "null"]
          format: date
      required:
        - department_id
        - content_type
        - headline
        - body
    UpdateSubmissionStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, published, needs_revision]
        admin_notes:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
      required:
        - status
...